+++ libraries/muxer/src/test/java/androidx/media3/muxer/FragmentedMp4MuxerEndToEndParameterizedTest.java
@@ -63,6 +63,7 @@ public class FragmentedMp4MuxerEndToEndParameterizedTest {
   private static final String AMR_WB_3GP = "3gp/bbb_mono_16kHz_23.05kbps_amrwb.3gp";
   private static final String OPUS_OGG = "ogg/bbb_6ch_8kHz_opus.ogg";
   private static final String VORBIS_OGG = "ogg/bbb_1ch_16kHz_q10_vorbis.ogg";
+  private static final String IAMF_OPUS_MP4 = "mp4/5p1_iamf_opus.mp4";
   private static final String RAW_WAV = "wav/bbb_2ch_44kHz.wav";
 
   public static final String MEDIA_ASSET_DIRECTORY = "asset:///media/";
@@ -85,6 +86,7 @@ public class FragmentedMp4MuxerEndToEndParameterizedTest {
         VP9_MP4,
         VP9_WEB,
         AMR_NB_3GP,
+        IAMF_OPUS_MP4,
         AMR_WB_3GP,
         OPUS_OGG,
         VORBIS_OGG,